splitted core in 3 different files
